backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
google-auth
pydantic
pydantic-settings
python-multipart
